+++ ThirdPartyAdapters/inmobi/CHANGELOG.md
@@ -1,7 +1,15 @@
 ## InMobi Android Mediation Adapter Changelog
 
+#### Version 10.0.9.0
+- Verified compatibility with InMobi SDK version 10.0.9.
+
+Built and tested with:
+- Google Mobile Ads SDK version 20.6.0.
+- InMobi SDK version 10.0.9.
+
 #### Version 10.0.8.0
 - Verified compatibility with InMobi SDK version 10.0.8.
+- Added Support for callback - onAdImpression
 
 Built and tested with:
 - Google Mobile Ads SDK version 20.6.0.